feat: create <TitleHeadingAndChildren />
